% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{create_hash}
\alias{create_hash}
\title{Create podcastindex authentication token}
\usage{
create_hash(epoch_time = NULL)
}
\arguments{
\item{epoch_time}{string of time in UTC unix format. If NULL, the current
system time will be used.}
}
\value{
string representing hash (sha1) of API key, secret and epoch time
created with the \code{digest} package
}
\description{
Create a custom authentication token required by the podcastindex api
utilizing the podcastindex API key and secret pair
}
