% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodes_requests.R
\name{episodes_byguid}
\alias{episodes_byguid}
\title{Get episodes from podcast guid}
\usage{
episodes_byguid(guid, feedurl = NULL, feedid = NULL, fulltext = FALSE)
}
\arguments{
\item{guid}{character string of guid value for episode to retrieve}

\item{feedurl}{character string of podcast feed URL}

\item{feedid}{vector of one or more integers representing PodcastIndex
Feed IDs.}

\item{fulltext}{boolean to return the full text value of any text fields
within the result data. If set to FALSE, all text fields are truncated
to 100 words.}
}
\value{
\code{tibble} data frame with metadata associated with the episode
}
\description{
\code{episodes_byguid()} obtains metadata for a single episode from a podcast
specified by a guid and either a feed ID or URL
}
\examples{
\dontshow{if (podindexr::podcastindex_api_isset()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Requires API key and secret

# using guid and feedid

episodes_byguid(guid = "PC2084", feedid = 920666)

# using guid and feedurl
episodes_byguid(guid = "PC2084", feedurl = "http://mp3s.nashownotes.com/pc20rss.xml")
\dontshow{\}) # examplesIf}
}
