% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodes_requests.R
\name{episodes_byid}
\alias{episodes_byid}
\title{Get metadata for a single episode}
\usage{
episodes_byid(id, fulltext = FALSE)
}
\arguments{
\item{id}{integer of the PodcastIndex episode ID to search for.}

\item{fulltext}{boolean to return the full text value of any text fields
within the result data. If set to FALSE, all text fields are truncated
to 100 words.}
}
\value{
\code{tibble} data frame with metadata associated with the episode
}
\description{
\code{episodes_byid()} obtains all metadata associated with a single episode
specified by ID.
}
\examples{
\dontshow{if (podindexr::podcastindex_api_isset()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Requires API key and secret

episodes_byid(id = 75075)
\dontshow{\}) # examplesIf}
}
