% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodes_requests.R
\name{episodes_byurl}
\alias{episodes_byurl}
\title{Get episodes from podcast feed URL}
\usage{
episodes_byurl(url, max = 10, fulltext = FALSE, since = NULL)
}
\arguments{
\item{url}{character string of podcast feed URL}

\item{max}{Maximum number of results to return.
Value must be a single integer between 1 and 1000.}

\item{fulltext}{boolean to return the full text value of any text fields
within the result data. If set to FALSE, all text fields are truncated
to 100 words.}

\item{since}{optional date string in format \code{YYYY-MM-DD} that when specified
will only return podcasts/episodes created since the specified date.}
}
\value{
\code{tibble} data frame with metadata associated with episodes
}
\description{
\code{episodes_byfeedid()} obtains all episodes from the a podcast specified by
a feed URL
}
\examples{
\dontshow{if (podindexr::podcastindex_api_isset()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Requires API key and secret

episodes_byurl(url = "https://engineered.network/pragmatic/feed/index.xml")
\dontshow{\}) # examplesIf}
}
