% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodes_requests.R
\name{episodes_live}
\alias{episodes_live}
\title{Get live podcast episodes}
\usage{
episodes_live(max = 10)
}
\arguments{
\item{max}{Maximum number of results to return.
Value must be a single integer between 1 and 1000.}
}
\value{
\code{tibble} data frame with metadata associated with live episodes
}
\description{
\code{episodes_live()} obtains metadata for episodes that have the
podcast-liveitem tag enabled in the feed, meaning episodes airing
live at the time of running this function.
}
\examples{
\dontshow{if (podindexr::podcastindex_api_isset()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Requires API key and secret

episodes_live(max = 20)
\dontshow{\}) # examplesIf}
}
