% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodes_requests.R
\name{episodes_random}
\alias{episodes_random}
\title{Random podcast episodes}
\usage{
episodes_random(
  max = 2,
  lang = "en",
  cat_terms = NULL,
  notcat_terms = NULL,
  fulltext = FALSE
)
}
\arguments{
\item{max}{Maximum number of results to return.
Value must be a single integer between 1 and 1000.}

\item{lang}{optional two-letter language code that when specified will
only return podcasts/episodes having that specific language.}

\item{cat_terms}{optional character vector of one or more categories that
when specified will only return podcasts/episodes under those categories.}

\item{notcat_terms}{optional character vector of one or more categories that
when specified will only return podcasts/episodes not under those
categories.}
}
\value{
\code{tibble} data frame with metadata associated with episodes
}
\description{
\code{episodes_random()} obtains a random selection of podcast episodes
from the Podcast Index database
}
\examples{
\dontshow{if (podindexr::podcastindex_api_isset()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Requires API key and secret

episodes_random(max = 20)
\dontshow{\}) # examplesIf}
}
