% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodes_requests.R
\name{episodes_recent}
\alias{episodes_recent}
\title{Get recent podcast episodes}
\usage{
episodes_recent(
  max = 7,
  exclude_string = NULL,
  before = NULL,
  fulltext = FALSE
)
}
\arguments{
\item{max}{Maximum number of results to return.
Value must be a single integer between 1 and 1000.}

\item{exclude_string}{character string used to exclude any episodes
with title or URL containing the specified string.}

\item{before}{episode ID string that when specified will only return
episodes created before the specified episode ID, allowing you to walk back
through episode history sequentially.}
}
\value{
\code{tibble} data frame with metadata associated with episodes
}
\description{
\code{episodes_recent()} obtains the most recent \code{max} number of podcast
episodes across the entire PodcastIndex database, sorted in reverse
chronological order.
}
\examples{
\dontshow{if (podindexr::podcastindex_api_isset()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Requires API key and secret

episodes_recent(max = 20)
\dontshow{\}) # examplesIf}
}
