% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeds_requests.R
\name{feeds_bymedium}
\alias{feeds_bymedium}
\title{Get podcast feeds by medium}
\usage{
feeds_bymedium(
  medium = c("audiobook", "blog", "film", "music", "newsletter", "podcast", "video")
)
}
\arguments{
\item{medium}{character string of medium to search for. Must be one of
\code{audiobook}, \code{blog}, \code{film}, \code{music}, \code{newsletter}, \verb{podcast,} or \code{video}}
}
\value{
\code{tibble} data frame with metadata associated with feeds
}
\description{
\code{feeds_bymedium} obtains all feeds marked with the specified medium tag
value.
}
\examples{
\dontrun{
# Requires API key and secret

feeds_bymedium(medium = "film")
}
}
