% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeds_requests.R
\name{feeds_recent}
\alias{feeds_recent}
\title{Get recent podcast feeds}
\usage{
feeds_recent(
  max = 40,
  lang = "en",
  since = NULL,
  cat_terms = NULL,
  notcat_terms = NULL
)
}
\arguments{
\item{max}{Maximum number of results to return.
Value must be a single integer between 1 and 1000.}

\item{lang}{optional two-letter language code that when specified will
only return podcasts/episodes having that specific language.}

\item{since}{optional date string in format \code{YYYY-MM-DD} that when specified
will only return podcasts/episodes created since the specified date.}

\item{cat_terms}{optional character vector of one or more categories that
when specified will only return podcasts/episodes under those categories.}

\item{notcat_terms}{optional character vector of one or more categories that
when specified will only return podcasts/episodes not under those
categories.}
}
\value{
\code{tibble} data frame with metadata associated with recent feeds
in chronological order
}
\description{
\code{feeds_recent()} retrieves the most recent podcast feeds added
to the podcastindex.
}
\examples{
\dontshow{if (podindexr::podcastindex_api_isset()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Requires API key and secret

feeds_recent()

feeds_recent(max = 100, since = "2021-01-01", cat_terms = c("News", "Finance"))
\dontshow{\}) # examplesIf}
}
