% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeds_requests.R
\name{feeds_withvalue}
\alias{feeds_withvalue}
\title{Get podcast feeds supporting value tag}
\usage{
feeds_withvalue()
}
\value{
\code{tibble} data frame with metadata associated with feeds
}
\description{
\code{feeds_withvalue} obtains all feeds from the Podcast Index database that
support the \code{value} tag in the podcast namespace
}
\examples{
\dontrun{
# Requires API key and secret

feeds_withvalue()
}
}
