% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{get_podcastindex_api_keys}
\alias{get_podcastindex_api_keys}
\title{Obtain podcastindex API key and secret pair}
\usage{
get_podcastindex_api_keys()
}
\value{
list with API key and API secret values
}
\description{
Obtain the podcastindex API key and secret values from a \code{.Renviron}
file, either stored in the current directory or the user's home directory.
}
