% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{podcastindex_api_isset}
\alias{podcastindex_api_isset}
\title{Detect if podcastindex API key/secret pair is defined}
\usage{
podcastindex_api_isset()
}
\value{
boolean, TRUE if both \code{PODCASTINDEX_API_KEY} and
\code{PODCASTINDEX_API_SECRET} environment variables are defined
in the current R session.
}
\description{
Detect if podcastindex API key/secret pair is defined
}
