% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podcasts_requests.R
\name{podcasts_byfeedid}
\alias{podcasts_byfeedid}
\title{Get podcast feed metadata}
\usage{
podcasts_byfeedid(feedid)
}
\arguments{
\item{feedid}{integer representing a Podcast Index feed ID}
}
\value{
\code{tibble} data frame with metadata associated with feed
}
\description{
\code{podcasts_byfeedid()} obtains all available metadata from the Podcast
Index database associated with a particular feed ID
}
\examples{
\dontshow{if (podindexr::podcastindex_api_isset()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Requires API key and secret

podcasts_byfeedid(feedid = 920666)
\dontshow{\}) # examplesIf}
}
