% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podcasts_requests.R
\name{podcasts_byfeedurl}
\alias{podcasts_byfeedurl}
\title{Get podcast feed metadata}
\usage{
podcasts_byfeedurl(url)
}
\arguments{
\item{url}{character string of podcast feed URL}
}
\value{
\code{tibble} data frame with metadata associated with feed
}
\description{
\code{podcasts_byfeedurl()} obtains all available metadata from the Podcast
Index database associated with a particular feed URL
}
\examples{
\dontshow{if (podindexr::podcastindex_api_isset()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Requires API key and secret

podcasts_byfeedurl(url = "https://feeds.theincomparable.com/batmanuniversity")
\dontshow{\}) # examplesIf}
}
