% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podcasts_requests.R
\name{podcasts_byguid}
\alias{podcasts_byguid}
\title{Get podcast feed metadata}
\usage{
podcasts_byguid(guid)
}
\arguments{
\item{guid}{string representing a Podcast Index feed GUID}
}
\value{
\code{tibble} data frame with metadata associated with feed
}
\description{
\code{podcasts_byfeedguid()} obtains all available metadata from the Podcast
Index database associated with a particular feed GUID
}
\examples{
\dontshow{if (podindexr::podcastindex_api_isset()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Requires API key and secret

podcasts_byguid(guid = "9b024349-ccf0-5f69-a609-6b82873eab3c")
\dontshow{\}) # examplesIf}
}
