% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\name{podcasts_categories}
\alias{podcasts_categories}
\title{Get podcast categories}
\usage{
podcasts_categories()
}
\value{
\code{tibble} data frame with category id and name
}
\description{
\code{podcasts_categories()} obtains all of the possible categories supported
by the Podcast Index.
}
\examples{
\dontshow{if (podindexr::podcastindex_api_isset()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Requires API key and secret

podcasts_categories()
\dontshow{\}) # examplesIf}
}
