% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_requests.R
\name{search_apple_lookup}
\alias{search_apple_lookup}
\title{Lookup podcasts with Apple-like API}
\usage{
search_apple_lookup(feedid)
}
\arguments{
\item{feedid}{integer of the iTunes feed ID to search for.}
}
\value{
\code{tibble} data frame with metadata associated with podcasts
that match the feed ID
}
\description{
\code{search_apple_lookup()} replaces the Apple podcast lookup api
but uses the Podcast Index database instead.
}
\examples{

search_apple_lookup(feedid = 1441923632)
}
