% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_requests.R
\name{search_bytitle}
\alias{search_bytitle}
\title{Search podcasts by title}
\usage{
search_bytitle(q, val = NULL, clean = TRUE, fulltext = FALSE, similar = FALSE)
}
\arguments{
\item{q}{character string of the term(s) to search for}

\item{val}{string indicating which value block type to filter
search results. Valid entries are \code{any}, \code{lightning}, \code{hive},
or \code{webmonetization}.}

\item{clean}{boolean to only return non-explicit podcast feeds. Meaning
feeds where the \code{itunes:explicit} flag is set to false.}

\item{fulltext}{boolean to return the full text value of any text fields
within the result data. If set to FALSE, all text fields are truncated
to 100 words.}

\item{similar}{boolean to include similar matches in search response}
}
\value{
\code{tibble} data frame with metadata associated with feeds
that match (or similar to) the search terms in the \code{title} fields.
}
\description{
\code{search_byterm()} performs a search of the \code{title} field of podcasts
feeds within the PodcastIndex.
}
\examples{
\dontshow{if (podindexr::podcastindex_api_isset()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Requires API key and secret

search_bytitle(q = "batman university")
\dontshow{\}) # examplesIf}
}
